Make Java version check <9 compatible
Make the Jave version check in `WorldFileData#getRegionData(CoordXZ)`
compatible with Java <9.  Remove the import of the
`java.lang.Runtime.Version` class (which was implemented in Java 9) from
`WorldFileData.java`.  Rewrite the comment describing the code block
handling the version check and execution of the `clear()` method of the
`header` `ByteBuffer` after type casting to `Buffer` or of the `header`
`ByteBuffer` (without type casting) based upon whether the Java version
being used is <9 or >=9, respectively.